refactor code

+++ lab1/gen2.py
@@ -15,25 +15,27 @@ class NaiveBayes:
         self.classes = None
         self.dataframe = None
         self.target = None
+        self.feature_columns = None
 
     def fit(self, dataframe, target):
         self.dataframe = dataframe
         self.target = target
         self.classes = self.dataframe[self.target].unique()
+        self.feature_columns = [col for col in dataframe.columns if col != self.target]
+        self.classes_prob = self.__calc_classes_prob()
+        self.features_prob_in_class_with_value = self.__calc_features_prob_in_class_with_value()
 
-        self.classes_prob = self._calc_classes_prob(dataframe, target)
-        self.features_prob_in_class_with_value = self._calc_features_prob_in_class_with_value(dataframe, target)
-
-    def predict(self, examples):
-        transform_ex = self._create_test_examples(examples, self.target)
+    def predict(self, test_examples):
+        transformed_ex = self.__create_test_examples(test_examples)
         predictions = []
-        for ex in transform_ex:
+
+        for ex in transformed_ex:
             prob = {}
             for cls in self.classes:
                 cls_prob = self.classes_prob[cls]
                 for feature, value in ex.items():
-                    if feature in self.features_prob_in_class_with_value and value in \
-                            self.features_prob_in_class_with_value[feature]:
+                    if (feature in self.features_prob_in_class_with_value and
+                            value in self.features_prob_in_class_with_value[feature]):
                         feature_in_class_prob = self.features_prob_in_class_with_value[feature][value].get(cls)
                         cls_prob *= feature_in_class_prob
                 prob[cls] = cls_prob
@@ -43,37 +45,31 @@ class NaiveBayes:
 
         return predictions
 
-    @staticmethod
-    def _create_test_examples(dataframe, target):
-        feature_columns = [col for col in dataframe.columns if col != target]
-        return [dict(zip(feature_columns, example)) for example in dataframe[feature_columns].values]
+    def __create_test_examples(self, test_array):
+        test_dataframe = pd.DataFrame(test_array, columns=self.feature_columns)
+        return [dict(zip(self.feature_columns, example)) for example in test_dataframe[self.feature_columns].values]
 
-    @staticmethod
-    def _calc_classes_prob(dataframe, target):
+    def __calc_classes_prob(self):
         prob = {}
-        total_count = len(dataframe)
-        for target_class in dataframe[target].unique():
-            prob[target_class] = len(
-                dataframe[dataframe[
-                              target] == target_class]) / total_count  # comment division to use count instead of probability
+        total_count = len(self.dataframe)
+
+        for cls in self.classes:
+            prob[cls] = len(self.dataframe[self.dataframe[self.target] == cls]) / total_count
 
         return prob
 
-    @staticmethod
-    def _calc_features_prob_in_class_with_value(dataframe, target):
+    def __calc_features_prob_in_class_with_value(self):
         prob = {}
-        feature_columns = [col for col in dataframe.columns if col != target]
 
-        for feature in feature_columns:
+        for feature in self.feature_columns:
             prob[feature] = {}
-            for value in dataframe[feature].unique():
+            for value in self.dataframe[feature].unique():
                 prob[feature][value] = {}
-                for target_class in dataframe[target].unique():
+                for cls in self.classes:
                     count_of_feature_in_class_with_value = len(
-                        dataframe[(dataframe[target] == target_class) & (dataframe[feature] == value)])
-                    count_of_feature_in_class = len(dataframe[dataframe[target] == target_class])
-                    prob[feature][value][
-                        target_class] = count_of_feature_in_class_with_value / count_of_feature_in_class  # comment division to use count instead of probability
+                        self.dataframe[(self.dataframe[self.target] == cls) & (self.dataframe[feature] == value)])
+                    count_of_feature_in_class = len(self.dataframe[self.dataframe[self.target] == cls])
+                    prob[feature][value][cls] = count_of_feature_in_class_with_value / count_of_feature_in_class
 
         return prob
 
@@ -81,32 +77,36 @@ class NaiveBayes:
 def separate_dataset(dataframe, target):
     x = dataframe.drop(columns=[target]).values
     y = dataframe[target].values
+
     return x, y
 
 
-def visualize_categorical_data(dataframe, hue_column=None, save_path='plots', show_plot=False):
-    if os.path.exists(save_path):
-        shutil.rmtree(save_path)
-    os.makedirs(save_path)
+def visualize_categorical_data(dataframe, target_column='class', save_path='plots', show_plot=False, save_plots=True):
+    if save_plots:
+        if os.path.exists(save_path):
+            shutil.rmtree(save_path)
+        os.makedirs(save_path)
 
     for i, col in enumerate(dataframe.columns):
-        if col == hue_column:
+        if col == target_column:
             continue
 
         plt.figure(figsize=(8, 6))
-        sns.countplot(x=col, hue=hue_column, data=dataframe)
+        sns.countplot(x=col, hue=target_column, data=dataframe)
         plt.title(col.replace('-', ' ').title())
-        plt.xlabel('')
+        plt.xlabel('Value')
         plt.ylabel('Count')
+        plt.legend(title=target_column, loc='best')
         plt.tight_layout()
 
-        filename = f"{save_path}/{col.replace(' ', '_').replace('-', '_')}.png"
-        plt.savefig(filename)
+        if save_plots:
+            filename = f"{save_path}/{col.replace(' ', '_').replace('-', '_')}.png"
+            plt.savefig(filename)
+            print(f"Plot saved at: {filename}")
 
         if show_plot:
             plt.show()
         plt.close()
-
 
 
 # cols = [
@@ -121,17 +121,17 @@ def visualize_categorical_data(dataframe, hue_column=None, save_path='plots', sh
 # df = pd.read_csv('datasets/agaricus-lepiota.data', names=cols)
 # target_feature = 'class'
 #
-cols = [
-    'class', 'handicapped-infants', 'water-project-cost-sharing',
-    'adoption-of-the-budget-resolution', 'physician-fee-freeze',
-    'el-salvador-aid', 'religious-groups-in-schools', 'anti-satellite-test-ban',
-    'aid-to-nicaraguan-contras', 'mx-missile', 'immigration',
-    'synfuels-corporation-cutback', 'education-spending', 'superfund-right-to-sue',
-    'crime', 'duty-free-exports', 'export-administration-act-south-africa'
-]
-
-df = pd.read_csv('datasets/house-votes-84.data', names=cols)
-target_feature = 'class'
+# cols = [
+#     'class', 'handicapped-infants', 'water-project-cost-sharing',
+#     'adoption-of-the-budget-resolution', 'physician-fee-freeze',
+#     'el-salvador-aid', 'religious-groups-in-schools', 'anti-satellite-test-ban',
+#     'aid-to-nicaraguan-contras', 'mx-missile', 'immigration',
+#     'synfuels-corporation-cutback', 'education-spending', 'superfund-right-to-sue',
+#     'crime', 'duty-free-exports', 'export-administration-act-south-africa'
+# ]
+#
+# df = pd.read_csv('datasets/house-votes-84.data', names=cols)
+# target_feature = 'class'
 #
 # cols = [
 #     'a1', 'a2', 'a3', 'a4', 'a5', 'a6',
@@ -170,26 +170,26 @@ target_feature = 'class'
 # df = pd.read_csv('datasets/breast-cancer.data', names=cols)
 # target_feature = 'class'
 #
-# cols = [
-#     'class',
-#     'date', 'plant-stand', 'precip', 'temp', 'hail', 'crop-hist', 'area-damaged',
-#     'severity', 'seed-tmt', 'germination', 'plant-growth', 'leaves', 'leafspots-halo',
-#     'leafspots-marg', 'leafspot-size', 'leaf-shread', 'leaf-malf', 'leaf-mild', 'stem',
-#     'lodging', 'stem-cankers', 'canker-lesion', 'fruiting-bodies', 'external-decay',
-#     'mycelium', 'int-discolor', 'sclerotia', 'fruit-pods', 'fruit-spots', 'seed',
-#     'mold-growth', 'seed-discolor', 'seed-size', 'shriveling', 'roots'
-# ]
-#
-# df = pd.read_csv('datasets/soybean-large.data', names=cols)
-# target_feature = 'class'
+cols = [
+    'class',
+    'date', 'plant-stand', 'precip', 'temp', 'hail', 'crop-hist', 'area-damaged',
+    'severity', 'seed-tmt', 'germination', 'plant-growth', 'leaves', 'leafspots-halo',
+    'leafspots-marg', 'leafspot-size', 'leaf-shread', 'leaf-malf', 'leaf-mild', 'stem',
+    'lodging', 'stem-cankers', 'canker-lesion', 'fruiting-bodies', 'external-decay',
+    'mycelium', 'int-discolor', 'sclerotia', 'fruit-pods', 'fruit-spots', 'seed',
+    'mold-growth', 'seed-discolor', 'seed-size', 'shriveling', 'roots'
+]
 
-visualize_categorical_data(df, hue_column=target_feature)
+df = pd.read_csv('datasets/soybean-large.data', names=cols)
+target_feature = 'class'
+
+visualize_categorical_data(df, target_column=target_feature)
 
 df = df.replace('?', pd.NA)
 
 df = df.ffill(axis=0)
 
-train, test = np.split(df.sample(frac=1), [int(0.7 * len(df))])
+train, test = np.split(df.sample(frac=1, random_state=42), [int(0.7 * len(df))])
 
 train = pd.DataFrame(train, columns=cols)
 test = pd.DataFrame(test, columns=cols)
@@ -199,5 +199,5 @@ model.fit(train, target_feature)
 
 X_test, y_test = separate_dataset(test, target_feature)
 
-y_predictions = model.predict(test)
+y_predictions = model.predict(X_test)
 print(classification_report(y_test, y_predictions))
